HEEDLS-370: Extract more test helper methods into test helper
