+++ DigitalLearningSolutions.Data.Tests/Helpers/ResetPasswordTestHelpers.cs
@@ -2,12 +2,14 @@
 {
     using System.Collections.Generic;
     using System.Data.Common;
+    using System.Linq;
+    using System.Threading.Tasks;
     using Dapper;
     using DigitalLearningSolutions.Data.Models.DbModels;
+    using DigitalLearningSolutions.Data.Models.User;
 
     public static class ResetPasswordTestHelpers
     {
-
         public static IEnumerable<ResetPassword> GetResetPasswordById(this DbConnection connection, int resetPasswordId)
         {
             return connection.Query<ResetPassword>
@@ -17,5 +19,24 @@
             );
         }
 
+        public static async Task<int> GetResetPasswordIdByHashAsync(this DbConnection connection, string hash)
+        {
+            var resetPasswordId = (await connection.QueryAsync<int>
+            (
+                "SELECT Id FROM ResetPassword WHERE ResetPasswordHash = @Hash;",
+                new { Hash = hash }
+            )).Single();
+            return resetPasswordId;
+        }
+
+        public static async Task SetResetPasswordIdForUserAsync
+            (this DbConnection connection, UserReference user, int resetPasswordId)
+        {
+            await connection.ExecuteAsync
+            (
+                $"UPDATE {user.UserType.TableName} SET ResetPasswordId = @ResetPasswordId WHERE {user.UserType.IdColumnName} = @UserId;",
+                new { ResetPasswordId = resetPasswordId, UserId = user.Id }
+            );
+        }
     }
 }
